Move continuous/stereographic RoPE to experimental; drop the module
Per review: StereographicRotaryPositionEmbedding2D was used purely as a 2D RoPE
table builder, never as a stateful module, so the module wrapper added no value
in production nn. Drop the class and relocate the four functional helpers
(build_rope_cos_sin_1d_continuous, build_axial_rope_cos_sin_2d_continuous,
stereographic_projection, spherical_centroid) to a new experimental module
physicsnemo/experimental/nn/rope.py, giving the continuous/stereographic RoPE
API room to mature outside the stable surface. The production integer-grid
builders (build_rope_cos_sin_1d, build_axial_rope_cos_sin_2d, apply_rotary_pos_emb,
RotaryPositionEmbedding{1,2}D) stay in physicsnemo/nn/module/rope.py.

Rewire DiT3D and PixelDiT to call build_axial_rope_cos_sin_2d_continuous
directly (storing head_dim / rope_base) instead of self.rope.build_tables;
coords.py imports the geometry helpers from physicsnemo.experimental.nn.

Tests: move the continuous-builder / projection / centroid tests to
test/experimental/nn/test_rope.py and convert the valuable module-property
tests (relative-position invariance, theta wiring, bf16 dtype) to function-level
tests; drop the module-API and now-redundant tests. The production
apply_rotary_pos_emb tests (incl. rotation-direction) stay in
test/nn/module/test_rope.py. import-linter contract unchanged (no
production -> experimental import).

This supersedes the earlier change that made the class a physicsnemo.Module
(MOD-001): the class is removed entirely.
