+++ physicsnemo/experimental/nn/__init__.py
@@ -23,5 +23,21 @@ changes between releases.
 
 from .flare_attention import FLARE
 from .diffusion_unet_3d_blocks import UNetBlock3D, Conv3D, GroupNorm3D, UNetAttention3D
+from .rope import (
+    build_axial_rope_cos_sin_2d_continuous,
+    build_rope_cos_sin_1d_continuous,
+    spherical_centroid,
+    stereographic_projection,
+)
 
-__all__ = ["FLARE", "UNetBlock3D", "Conv3D", "GroupNorm3D", "UNetAttention3D"]
+__all__ = [
+    "FLARE",
+    "UNetBlock3D",
+    "Conv3D",
+    "GroupNorm3D",
+    "UNetAttention3D",
+    "build_axial_rope_cos_sin_2d_continuous",
+    "build_rope_cos_sin_1d_continuous",
+    "spherical_centroid",
+    "stereographic_projection",
+]
