+++ physicsnemo/nn/module/__init__.py
@@ -109,14 +109,9 @@ from .resample_layers import (
 from .rope import (
     RotaryPositionEmbedding1D,
     RotaryPositionEmbedding2D,
-    StereographicRotaryPositionEmbedding2D,
     apply_rotary_pos_emb,
     build_axial_rope_cos_sin_2d,
-    build_axial_rope_cos_sin_2d_continuous,
     build_rope_cos_sin_1d,
-    build_rope_cos_sin_1d_continuous,
-    spherical_centroid,
-    stereographic_projection,
 )
 from .siren_layers import SirenLayer, SirenLayerType
 from .spectral_layers import (
